refactor ui into submodules
